Cleanup: extract DBusIPCProxyBinder class
