+++ src/ipc/dbus/DBusIPCProxy.h
@@ -33,6 +33,7 @@
 #include "ipc-dbus.h"
 #include "IPCProxyBase.h"
 #include "DBusRequestHandler.h"
+#include "DBusIPCProxyBinder.h"
 #include <QDBusServiceWatcher>
 
 #if defined(FaceliftIPCLibDBus_LIBRARY)
@@ -48,98 +49,6 @@ namespace dbus {
 using namespace facelift;
 
 class DBusManager;
-
-class FaceliftIPCLibDBus_EXPORT DBusIPCProxyBinder : public IPCProxyBinderBase
-{
-    Q_OBJECT
-
-public:
-
-    template<typename Type>
-    using IPCProxyType = typename Type::IPCDBusProxyType;
-
-    DBusIPCProxyBinder(InterfaceBase &owner, QObject *parent = nullptr);
-
-    const QString &serviceName() const
-    {
-        return m_serviceName;
-    }
-
-    void setServiceName(const QString &name);
-
-    const QString &interfaceName() const
-    {
-        return m_interfaceName;
-    }
-
-    void setInterfaceName(const QString &name);
-
-    Q_SLOT void onPropertiesChanged(const QDBusMessage &dbusMessage);
-
-    Q_SLOT void onSignalTriggered(const QDBusMessage &dbusMessage);
-
-    void bindToIPC() override;
-
-    void onServiceNameKnown();
-
-    void setServiceAvailable(bool isRegistered);
-
-    bool isServiceAvailable() const override
-    {
-        return m_serviceAvailable;
-    }
-
-    void requestPropertyValues();
-
-    DBusIPCMessage call(DBusIPCMessage &message) const;
-
-    void asyncCall(DBusIPCMessage &message, const QObject *context, std::function<void(DBusIPCMessage &message)> callback);
-
-    template<typename Type>
-    void serializeValue(DBusIPCMessage &msg, const Type &v);
-
-    template<typename Type>
-    void deserializeValue(DBusIPCMessage &msg, Type &v);
-
-    void onServerNotAvailableError(const char *methodName) const;
-
-    template<typename PropertyType>
-    void sendSetterCall(const char *methodName, const PropertyType &value);
-
-    template<typename ... Args>
-    DBusIPCMessage sendMethodCall(const char *methodName, const Args & ... args) const;
-
-    template<typename ReturnType, typename ... Args>
-    void sendAsyncMethodCall(const char *methodName, facelift::AsyncAnswer<ReturnType> answer, const Args & ... args);
-
-    template<typename ... Args>
-    void sendAsyncMethodCall(const char *methodName, facelift::AsyncAnswer<void> answer, const Args & ... args);
-
-    template<typename ReturnType, typename ... Args>
-    void sendMethodCallWithReturn(const char *methodName, ReturnType &returnValue, const Args & ... args) const;
-
-    QDBusConnection &connection() const;
-
-    DBusManager &manager() const;
-
-    void setHandler(DBusRequestHandler *handler)
-    {
-        m_serviceObject = handler;
-        checkInit();
-    }
-
-private:
-    void checkRegistry();
-
-    QString m_serviceName;
-    QString m_interfaceName;
-    QDBusServiceWatcher m_busWatcher;
-    DBusRequestHandler *m_serviceObject = nullptr;
-    bool m_explicitServiceName = false;
-    bool m_serviceAvailable = false;
-    DBusObjectRegistry &m_registry;
-};
-
 
 class FaceliftIPCLibDBus_EXPORT DBusIPCProxyBase : protected DBusRequestHandler
 {
