+++ src/ipc/dbus/ipc-dbus-serialization.h
@@ -63,95 +63,7 @@ inline void IPCDBusServiceAdapterBase::deserializeValue(DBusIPCMessage &msg, Typ
 }
 
 
-template<typename Type>
-inline void DBusIPCProxyBinder::serializeValue(DBusIPCMessage &msg, const Type &v)
-{
-    typedef typename IPCTypeRegisterHandler<Type>::SerializedType SerializedType;
-    IPCTypeHandler<SerializedType>::write(msg.outputPayLoad(), IPCTypeRegisterHandler<Type>::convertToSerializedType(v, *this));
-}
 
-template<typename Type>
-inline void DBusIPCProxyBinder::deserializeValue(DBusIPCMessage &msg, Type &v)
-{
-    typedef typename IPCTypeRegisterHandler<Type>::SerializedType SerializedType;
-    SerializedType serializedValue;
-    IPCTypeHandler<SerializedType>::read(msg.inputPayLoad(), serializedValue);
-    IPCTypeRegisterHandler<Type>::convertToDeserializedType(v, serializedValue, *this);
-}
-
-
-template<typename ... Args>
-inline DBusIPCMessage DBusIPCProxyBinder::sendMethodCall(const char *methodName, const Args & ... args) const
-{
-    DBusIPCMessage msg(m_serviceName, objectPath(), m_interfaceName, methodName);
-    auto argTuple = std::make_tuple(args ...);
-    for_each_in_tuple(argTuple, SerializeParameterFunction<DBusIPCProxyBinder>(msg.outputPayLoad(), *this));
-    auto replyMessage = this->call(msg);
-    if (replyMessage.isErrorMessage()) {
-        onServerNotAvailableError(methodName);
-    }
-    return replyMessage;
-}
-
-template<typename ReturnType, typename ... Args>
-inline void DBusIPCProxyBinder::sendAsyncMethodCall(const char *methodName, facelift::AsyncAnswer<ReturnType> answer, const Args & ... args)
-{
-    DBusIPCMessage msg(m_serviceName, objectPath(), m_interfaceName, methodName);
-    auto argTuple = std::make_tuple(args ...);
-    for_each_in_tuple(argTuple, SerializeParameterFunction<DBusIPCProxyBinder>(msg.outputPayLoad(), *this));
-    asyncCall(msg, this, [this, answer](DBusIPCMessage &msg) {
-        ReturnType returnValue;
-        if (msg.isReplyMessage()) {
-            deserializeValue(msg, returnValue);
-            answer(returnValue);
-        } else {
-            qCWarning(LogIpc) << "Error received" << msg.toString();
-        }
-    });
-}
-
-template<typename ... Args>
-inline void DBusIPCProxyBinder::sendAsyncMethodCall(const char *methodName, facelift::AsyncAnswer<void> answer, const Args & ... args)
-{
-    DBusIPCMessage msg(m_serviceName, objectPath(), m_interfaceName, methodName);
-    auto argTuple = std::make_tuple(args ...);
-    for_each_in_tuple(argTuple, SerializeParameterFunction<DBusIPCProxyBinder>(msg.outputPayLoad(), *this));
-    asyncCall(msg, this, [answer](DBusIPCMessage &msg) {
-        Q_UNUSED(msg);
-        answer();
-    });
-}
-
-template<typename ReturnType, typename ... Args>
-inline void DBusIPCProxyBinder::sendMethodCallWithReturn(const char *methodName, ReturnType &returnValue, const Args & ... args) const
-{
-    DBusIPCMessage msg = sendMethodCall(methodName, args ...);
-    if (msg.isReplyMessage()) {
-        const_cast<DBusIPCProxyBinder *>(this)->deserializeValue(msg, returnValue);
-    } else {
-        assignDefaultValue(returnValue);
-    }
-}
-
-
-template<typename PropertyType>
-inline void DBusIPCProxyBinder::sendSetterCall(const char *methodName, const PropertyType &value)
-{
-    DBusIPCMessage msg(m_serviceName, objectPath(), m_interfaceName, methodName);
-    serializeValue(msg, value);
-    if (isSynchronous()) {
-        auto replyMessage = call(msg);
-        if (replyMessage.isErrorMessage()) {
-            onServerNotAvailableError(methodName);
-        }
-    } else {
-        asyncCall(msg, this, [this, methodName](const DBusIPCMessage &replyMessage) {
-            if (replyMessage.isErrorMessage()) {
-                onServerNotAvailableError(methodName);
-            }
-        });
-    }
-}
 
 
 template<typename MemberID, typename ... Args>
